# Speed up `OGIP` unit parser for composite units