--- a/docs/changes/units/16761.perf.rst
+++ b/docs/changes/units/16761.perf.rst
@@ -1,0 +1,1 @@
+Parsing composite units with the OGIP formatter is now up to 25% faster.